[DO_NOT_MERGE]: Fix MAINTAINERS file
There are people listed in this MAINTAINERS file which do not have
Gerrit IDs, or at least do not have Gerrit IDs matching the email
addresses listed in this file.  This causes problems for some people's
workflow tools which attempt to automatically add reviewers based on the
entries in the MAINTAINERS file.

Please update the email addresses or delete the MAINTAINERS.

Change-Id: I065c2db07a8bf9003db8239ccd9b43ce6f2c5e8e
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>